add cloudinary functionality in trusted by

+++ src/app/api/trusted-by/[id]/route.ts
@@ -1,13 +1,32 @@
 import prisma from "@/utils/prisma";
 import { NextResponse } from "next/server";
+import { v2 as cloudinary } from "cloudinary";
 
+cloudinary.config({
+  cloud_name: process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 export async function DELETE(
   request: Request,
   { params }: { params: { id: string } },
 ) {
   try {
     const { id } = params;
-    console.log(id);
+    const trustedBy = await prisma.trustedby.findUnique({
+      where: { id },
+    });
+
+    if (!trustedBy)
+      return NextResponse.json(
+        { message: "Trusted by not found" },
+        { status: 404 },
+      );
+
+    await cloudinary.api.delete_resources([trustedBy.image], {
+      type: "upload",
+      resource_type: "image",
+    });
     await prisma.trustedby.delete({ where: { id } });
     return NextResponse.json(
       { message: "Trusted By deleted" },
